Docker, new scripts, renaming and restructuring
