+++ .travis.yml
@@ -1,16 +1,16 @@
 sudo: required
 language: python
 python:
-  - 2.7
-group: stable
-dist: trusty
+  - 3.6
+
+services:
+  - docker
 
 install:
-  - pip install -r requirements.txt
+  - docker-compose up -d scraper
 
 script:
-  - cd FreeEyeOut/spiders
-  - scrapy runspider FreeShittyEyeOut.py
+  - bash scripts/live_tests.sh
 
 after_script:
   - coveralls
