+++ Dockerfile.scraper
@@ -1,0 +1,7 @@
+FROM python:3
+WORKDIR /home/FreeEyeOut
+
+# Install requirements
+COPY requirements.txt .
+RUN pip install -r requirements.txt
+COPY . .
