+++ docker-compose.yml
@@ -1,0 +1,13 @@
+version: '3'
+
+services:
+  scraper:
+    environment:
+    - PYTHONUNBUFFERED=1
+    build:
+      context: .
+      dockerfile: Dockerfile.scraper
+    volumes:
+      - .:/home/FreeEyeOut
+    ports:
+      - "8000:8080"
