+++ scraper/scrapy.cfg
@@ -4,8 +4,8 @@
 # https://scrapyd.readthedocs.org/en/latest/deploy.html
 
 [settings]
-default = FreeEyeOut.settings
+default = free_scraper.settings
 
 [deploy]
 #url = http://localhost:6800/
-project = FreeEyeOut
+project = free_scraper
