+++ scripts/live_tests.sh
@@ -1,0 +1,5 @@
+set -e
+echo "Running live crawler tests..."
+cd ./scraper/scraper && scrapy crawl course_spider
+echo "Test complete."
+set +e
